requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pre-commit==3.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
virtualenv==20.26.2
